Add files via upload
This project involves the analysis of two datasets related to Airbnb. The first dataset contains information on guests who are searching for accommodation on the platform, while the second dataset pertains to the inquiries that hosts tend to accept. The primary objective of this project is to conduct a comprehensive analysis of the market demand and supply dynamics, with a focus on data visualization and communication of key insights.

The findings of the project indicate several significant conclusions.
*Firstly, it was observed that the peak season for Airbnb tends to be from October to December. 
*Secondly, the most popular type of accommodation on the platform is the two-room private apartment, while there appears to be a shortage of one-room apartments as the number of searches for such accommodation is double that of inquiries for two-room apartments.
*Thirdly, it was noted that the percentage of inquiries accepted by hosts is only 50%, which suggests that the demand for accommodation may be larger than the supply. 
*Finally, the project utilized a random forest algorithm, which